backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
SQLAlchemy==2.0.36
alembic==1.13.3
asyncpg==0.30.0
psycopg2-binary==2.9.10
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.17
httpx==0.27.2
redis==5.2.0
celery==5.4.0
email-validator==2.2.0
python-dateutil==2.9.0.post0
tenacity==9.0.0
structlog==24.4.0
sentry-sdk[fastapi]==2.17.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
